% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brutforce knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{vector}

\item{W}{vector}

\item{parallel}{logical}
}
\value{
list
}
\description{
brutforce knapsack
}
