% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab6_Greedy.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Function Knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{is a dataframe containing 2 elements w and v.}

\item{W}{is the maximum weight that should be in a pack.}
}
\value{
Maximum value of all objects in the pack.
}
\description{
Greedy Function Knapsack
}
