% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_lab06.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{dynamic function for knapsack}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{a dataframe including v and w}

\item{W}{maximum weight}
}
\value{
matimum value and elements
}
\description{
dynamic function for knapsack
}
